requirements.txt
Flask==2.3.3
flask-cors==3.0.10
flasgger==0.10.5
requests==2.32.0
python-dotenv==1.0.0
google-genai==0.5.2
